Removed unused imports, move google icon, update sign in text
