+++ components/Comments/GoogleLoginButton.tsx
@@ -1,9 +1,8 @@
-import { CredentialResponse, GoogleLogin } from "@react-oauth/google";
-import { useState } from "react";
+import { GoogleLogin } from "@react-oauth/google";
 import { handleGoogleLogin } from "../../lib/api/api";
 import { useStore } from "../../lib/store/userStore";
 import { useMutation } from "@tanstack/react-query";
-import axios, { Axios, AxiosError } from "axios";
+import axios, { AxiosError } from "axios";
 import toast from "react-hot-toast";
 export default function GoogleLoginButton() {
   const [user, setUser] = useStore((state) => [state.user, state.setUser]);
@@ -22,7 +21,7 @@ export default function GoogleLoginButton() {
     onSuccess: (data) => {
       localStorage.setItem("user", JSON.stringify(data));
       setUser(data);
-      notify("success", "Login success!");
+      notify("success", "Signin successful!");
     },
     onError: (error: AxiosError | Error) => {
       if (axios.isAxiosError(error)) {
